Move extensions to MediaMetadataCompatExt and expose artwork image URI
