+++ common/src/main/java/com/example/android/uamp/media/MusicService.kt
@@ -273,29 +273,6 @@ open class MusicService : MediaBrowserServiceCompat(), SessionAvailabilityListen
             currentPlayer.playWhenReady = playWhenReady
         }
     }
-
-    private fun MediaMetadataCompat.toMediaQueueItem(): MediaQueueItem {
-        val metadata: MediaMetadata = toCastMediaMetadata()
-        val mediaInfo = MediaInfo.Builder(this.mediaUri.toString())
-                .setStreamType(MediaInfo.STREAM_TYPE_BUFFERED)
-                .setContentType(MimeTypes.AUDIO_MPEG)
-                .setStreamDuration(this.duration)
-                .setMetadata(metadata)
-                .build()
-        return MediaQueueItem.Builder(mediaInfo).build()
-    }
-
-    private fun MediaMetadataCompat.toCastMediaMetadata(): MediaMetadata {
-        val mediaMetadata = MediaMetadata(MediaMetadata.MEDIA_TYPE_MUSIC_TRACK)
-        mediaMetadata.putString(MediaMetadata.KEY_TITLE, this.title)
-        mediaMetadata.putString(MediaMetadata.KEY_ARTIST, this.artist)
-        mediaMetadata.putString(MediaMetadata.KEY_ALBUM_TITLE, this.album)
-        mediaMetadata.putString(MediaMetadata.KEY_ALBUM_ARTIST, this.albumArtist)
-        mediaMetadata.putString(MediaMetadata.KEY_COMPOSER, this.composer)
-        this.date?.let { date -> mediaMetadata.putString(MediaMetadata.KEY_RELEASE_DATE, date) }
-        mediaMetadata.putInt(MediaMetadata.KEY_TRACK_NUMBER, this.trackNumber.toInt())
-        mediaMetadata.putInt(MediaMetadata.KEY_DISC_NUMBER, this.discNumber.toInt())
-        return mediaMetadata
     }
 
     /**
